move tests to integration tests folder
