+++ squire_core/Cargo.toml
@@ -5,6 +5,8 @@ edition = "2021"
 license = "AGPL-3.0"
 authors = ["TylerBloom <tylerbloom2222@gmail.com>"]
 
+[features]
+db-tests = []
 
 [dependencies]
 # In-House deps
