+++ squire_core/src/tests/mod.rs
@@ -1,6 +1,7 @@
 mod init;
 mod requests;
 mod utils;
+mod state;
 
 //pub mod tournaments;
 //pub use tournaments::*;
